+++ CQ.Application/SystemManage/AreaApp.cs
@@ -24,7 +24,7 @@ namespace CQ.Application.SystemManage
         {
             return service.FindEntity(keyValue);
         }
-        public void DeleteForm(long keyValue)
+        public void DeleteForm(int keyValue)
         {
             if (service.IQueryable().Count(t => t.F_ParentId.Equals(keyValue)) > 0)
             {
@@ -35,7 +35,7 @@ namespace CQ.Application.SystemManage
                 service.Delete(t => t.F_Id == keyValue);
             }
         }
-        public void SubmitForm(AreaEntity areaEntity, long keyValue)
+        public void SubmitForm(AreaEntity areaEntity, int keyValue)
         {
             if (keyValue > 0)
             {
