+++ CQ.Domain/01 Infrastructure/ICreationAudited.cs
@@ -4,8 +4,8 @@ namespace CQ.Domain
 {
     public interface ICreationAudited
     {
-        long F_Id { get; set; }
-        long? F_CreatorUserId { get; set; }
+        int F_Id { get; set; }
+        int? F_CreatorUserId { get; set; }
         DateTime? F_CreatorTime { get; set; }
     }
 }
