+++ CQ.Domain/01 Infrastructure/IModificationAudited.cs
@@ -4,8 +4,8 @@ namespace CQ.Domain
 {
     public interface IModificationAudited
     {
-        long F_Id { get; set; }
-        long? F_LastModifyUserId { get; set; }
+        int F_Id { get; set; }
+        int? F_LastModifyUserId { get; set; }
         DateTime? F_LastModifyTime { get; set; }
     }
 }
