+++ CQ.Domain/03 Entity/BusinessData/ProductEntity.cs
@@ -9,7 +9,7 @@ namespace CQ.Domain.Entity.BusinessData
 {
     public class ProductEntity : IEntity<ProductEntity>, ICreationAudited, IDeleteAudited, IModificationAudited
     {
-        public long F_Id { get; set; }
+        public int F_Id { get; set; }
         public string F_ProductName { get; set; }
         public bool F_IsHot { get; set; }
         public string F_ProductImg { get; set; }
@@ -22,9 +22,9 @@ namespace CQ.Domain.Entity.BusinessData
         public DateTime? F_CreatorTime { get; set; }
         public int? F_CreatorUserId { get; set; }
         public DateTime? F_LastModifyTime { get; set; }
-        public long? F_LastModifyUserId { get; set; }
+        public int? F_LastModifyUserId { get; set; }
         public DateTime? F_DeleteTime { get; set; }
-        public long? F_DeleteUserId { get; set; }
+        public int? F_DeleteUserId { get; set; }
         public string F_Remark { get; set; }
     }
 }
