+++ CQ.Domain/03 Entity/SystemManage/ItemsEntity.cs
@@ -9,8 +9,8 @@ namespace CQ.Domain.Entity.SystemManage
 {
     public class ItemsEntity : IEntity<ItemsEntity>, ICreationAudited, IDeleteAudited, IModificationAudited
     {
-        public long F_Id { get; set; }
-        public long F_ParentId { get; set; }
+        public int F_Id { get; set; }
+        public int F_ParentId { get; set; }
         public string F_EnCode { get; set; }
         public string F_FullName { get; set; }
         public bool? F_IsTree { get; set; }
@@ -20,10 +20,10 @@ namespace CQ.Domain.Entity.SystemManage
         public bool? F_EnabledMark { get; set; }
         public string F_Description { get; set; }
         public DateTime? F_CreatorTime { get; set; }
-        public long? F_CreatorUserId { get; set; }
+        public int? F_CreatorUserId { get; set; }
         public DateTime? F_LastModifyTime { get; set; }
-        public long? F_LastModifyUserId { get; set; }
+        public int? F_LastModifyUserId { get; set; }
         public DateTime? F_DeleteTime { get; set; }
-        public long? F_DeleteUserId { get; set; }
+        public int? F_DeleteUserId { get; set; }
     }
 }
