+++ CQ.Domain/03 Entity/SystemManage/RoleAuthorizeEntity.cs
@@ -9,13 +9,13 @@ namespace CQ.Domain.Entity.SystemManage
 {
     public class RoleAuthorizeEntity : IEntity<RoleAuthorizeEntity>, ICreationAudited
     {
-        public long F_Id { get; set; }
+        public int F_Id { get; set; }
         public int? F_ItemType { get; set; }
-        public long F_ItemId { get; set; }
+        public int F_ItemId { get; set; }
         public int? F_ObjectType { get; set; }
-        public long F_ObjectId { get; set; }
+        public int F_ObjectId { get; set; }
         public int? F_SortCode { get; set; }
         public DateTime? F_CreatorTime { get; set; }
-        public long? F_CreatorUserId { get; set; }
+        public int? F_CreatorUserId { get; set; }
     }
 }
