+++ CQ.Domain/04 IRepository/SystemManage/IUserRepository.cs
@@ -5,7 +5,7 @@ namespace CQ.Domain.IRepository.SystemManage
 {
     public interface IUserRepository : IRepositoryBase<UserEntity>
     {
-        void DeleteForm(long keyValue);
-        void SubmitForm(UserEntity userEntity, UserLogOnEntity userLogOnEntity, long keyValue);
+        void DeleteForm(int keyValue);
+        void SubmitForm(UserEntity userEntity, UserLogOnEntity userLogOnEntity, int keyValue);
     }
 }
