+++ CQ.Permission/Areas/ContentManage/Controllers/GamesController.cs
@@ -31,7 +31,7 @@ namespace CQ.Permission.Areas.ContentManage.Controllers
         }
         [HttpGet]
         [HandlerAjaxOnly]
-        public ActionResult GetFormJson(long keyValue)
+        public ActionResult GetFormJson(int keyValue)
         {
             var data = productApp.GetForm(keyValue);
             return Content(data.ToJson());
@@ -41,7 +41,7 @@ namespace CQ.Permission.Areas.ContentManage.Controllers
         [HandlerAjaxOnly]
         [ValidateAntiForgeryToken]
         [ValidateInput(false)]
-        public ActionResult SubmitForm(ProductEntity productEntity, long keyValue)
+        public ActionResult SubmitForm(ProductEntity productEntity, int keyValue)
         {
             string imgListStr = productEntity.F_Remark;
             productApp.SubmitForm(productEntity, imgListStr.TrimEnd(',').Split(','), keyValue);
@@ -52,7 +52,7 @@ namespace CQ.Permission.Areas.ContentManage.Controllers
         [HandlerAuthorize]
         [HandlerAjaxOnly]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteForm(long keyValue)
+        public ActionResult DeleteForm(int keyValue)
         {
             productApp.DeleteForm(keyValue);
             return Success("删除成功。");
