+++ CQ.Permission/Areas/SystemManage/Controllers/AreaController.cs
@@ -63,7 +63,7 @@ namespace CQ.Permission.Areas.SystemManage.Controllers
         [HttpPost]
         [HandlerAjaxOnly]
         [ValidateAntiForgeryToken]
-        public ActionResult SubmitForm(AreaEntity areaEntity, long keyValue)
+        public ActionResult SubmitForm(AreaEntity areaEntity, int keyValue)
         {
             areaApp.SubmitForm(areaEntity, keyValue);
             return Success("操作成功。");
@@ -72,7 +72,7 @@ namespace CQ.Permission.Areas.SystemManage.Controllers
         [HandlerAjaxOnly]
         [HandlerAuthorize]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteForm(long keyValue)
+        public ActionResult DeleteForm(int keyValue)
         {
             areaApp.DeleteForm(keyValue);
             return Success("删除成功。");
